requirements.txt
Django
djangorestframework
psycopg2-binary==2.9.6
gunicorn==20.1.0
python-dotenv==1.0.0
django-phonenumber-field==7.1.0
phonenumbers==8.13.11